+++ lib/src/client/cargo.dart
@@ -3,7 +3,7 @@ part of cargo_client;
 /// Cargo storage backends for client
 const String CARGO_MODE_LOCAL = "localstorage";
 
-abstract class Cargo extends CargoBase with CargoDispatch {
+abstract class Cargo extends CargoBase {
   Cargo._();
   /// Create a new cargo storage
   factory Cargo({CargoModeHolder MODE: CargoMode.MEMORY, Map path}) {