+++ lib/src/shared/cargo_dispatch.dart
@@ -3,14 +3,17 @@ part of cargo;
 typedef DataChangeListener(DataEvent de);
 
 class DataEvent {
+  dynamic key;
   dynamic data;
   
-  DataEvent(this.data);
+  DataEvent(this.key, this.data);
 }
 
 class CargoDispatch {
   
   Map<String, List<DataChangeListener>> mapping = new Map<String, List<DataChangeListener>>();
+  
+  DataChangeListener _cargoDataChange;
   
   on(String key, DataChangeListener cargoDataChange) {
     List<DataChangeListener> cargoDataChangerList = new List<DataChangeListener>();
@@ -21,6 +24,8 @@ class CargoDispatch {
     }
     cargoDataChangerList.add(cargoDataChange);
   }
+  
+  onAll(DataChangeListener cargoDataChange) => this._cargoDataChange = cargoDataChange;
   
   off(String key, DataChangeListener cargoDataChange) {
     if (mapping[key]!=null) {
@@ -36,9 +41,11 @@ class CargoDispatch {
     if (mapping[key]!=null) {
       List<DataChangeListener> cargoDataChangerList = mapping[key];
       
+      DataEvent dataEvent = new DataEvent(key, value);
       for (DataChangeListener dataChangeListener in cargoDataChangerList) {
-        dataChangeListener(new DataEvent(value));
+        dataChangeListener(dataEvent);
       }
+      if (this._cargoDataChange!=null) this._cargoDataChange(dataEvent);
     }
   }