+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file contains highlights of what changes on each version of the cargo package.
 
+#### Pub version 0.4.4 ####
+
+- Add the key to the dataEvent object.
+- Add onAll method to CargoDispatcher
+
 #### Pub version 0.4.3 ####
 
 - length method returns now Future<int> instead of int, go to an asynchronous implementation.